Fix integer conversion warnings in library and test suites

+++ include/mbedtls/asn1write.h
@@ -26,7 +26,7 @@
 #include "asn1.h"
 
 #define MBEDTLS_ASN1_CHK_ADD(g, f) do { if( ( ret = f ) < 0 ) return( ret ); else   \
-                                g += ret; } while( 0 )
+                                g += (size_t) ret; } while( 0 )
 
 #ifdef __cplusplus
 extern "C" {
