+++ library/md5.c
@@ -343,7 +343,7 @@ static const unsigned char md5_test_buf[7][81] =
       "345678901234567890" }
 };
 
-static const int md5_test_buflen[7] =
+static const size_t md5_test_buflen[7] =
 {
     0, 1, 3, 14, 26, 62, 80
 };
