+++ library/x509_crl.c
@@ -337,7 +337,7 @@ int mbedtls_x509_crl_parse_der( mbedtls_x509_crl *chain,
     }
 
     end = p + len;
-    crl->tbs.len = end - crl->tbs.p;
+    crl->tbs.len = (size_t) ( end - crl->tbs.p );
 
     /*
      * Version  ::=  INTEGER  OPTIONAL {  v1(0), v2(1)  }
@@ -386,7 +386,7 @@ int mbedtls_x509_crl_parse_der( mbedtls_x509_crl *chain,
         return( ret );
     }
 
-    crl->issuer_raw.len = p - crl->issuer_raw.p;
+    crl->issuer_raw.len = (size_t) ( p - crl->issuer_raw.p );
 
     /*
      * thisUpdate          Time
